requirements.txt
# ═══════════════════════════════════════════════════════════════
#                    MitaHelper - Зависимости
# ═══════════════════════════════════════════════════════════════

# Основная библиотека Telegram (устанавливает свои зависимости автоматически)
python-telegram-bot==13.15

# База данных (опционально)
# sqlalchemy==2.0.25
# psycopg2-binary==2.9.9
# pymongo==4.6.1

# Утилиты
python-dotenv==1.0.0
requests==2.31.0
